API/ServiceLogic/InvitationService.cs
using Domain;
using IRepository;
using ServiceLogic.CustomExceptions;

namespace ServiceLogic;

public class InvitationService
{

    private readonly IInvitationRepository _invitationRepository;
    public InvitationService(IInvitationRepository invitationRepository)
    {
        _invitationRepository = invitationRepository;
    }
    
    
    public IEnumerable<Invitation> GetAllInvitations()
    {

        try
        {
            IEnumerable<Invitation> invitations = _invitationRepository.GetAllInvitations();
            return invitations;
        }
        catch (Exception exceptionCaught)
        {
            throw new UnknownServiceException(exceptionCaught.Message);
        }
        
    }
}